Add template link to button so it deploy Azure resource
Before I had an <img> of the deploy button as a placemarker.
Replaced this HTML with Markdown and included a link to deploy
the resource. The link to deploy the resource consisted of a
base url pointing to Azure portal, and the url slug is an
encoded version of the raw url to my template in my repo.
Instructions to do this at:
https://docs.microsoft.com/en-us/azure/azure-resource-manager/templates/deploy-to-azure-button